+++ src/main/java/com/example/demo/repository/LabelRepository.java
@@ -2,8 +2,7 @@ package com.example.demo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.model.Brand;
 import com.example.demo.model.DemandLabel;
 
-public interface LabelRepository extends JpaRepository<DemandLabel, Long> {
+public interface LabelRepository extends JpaRepository<DemandLabel, Integer> {
 }
